fix(rust, python): fix null/empty in List::take_unchecked
