+++ polars/polars-core/src/chunked_array/kernels/take.rs
@@ -1,9 +1,8 @@
 use std::convert::TryFrom;
 
-use arrow::bitmap::MutableBitmap;
-use polars_arrow::array::PolarsArray;
-use polars_arrow::bit_util::unset_bit_raw;
+use polars_arrow::compute::take::bitmap::take_bitmap_unchecked;
 use polars_arrow::compute::take::take_value_indices_from_list;
+use polars_arrow::utils::combine_validities_and;
 
 use crate::prelude::*;
 
@@ -54,28 +53,13 @@ pub(crate) unsafe fn take_list_unchecked(
 
     let taken = taken.array_ref(0).clone();
 
-    let validity =
-        // if null count > 0
-        if values.has_validity() || indices.has_validity() {
-            // determine null buffer, which are a function of `values` and `indices`
-            let mut validity = MutableBitmap::with_capacity(indices.len());
-            let validity_ptr = validity.as_slice().as_ptr() as *mut u8;
-            validity.extend_constant(indices.len(), true);
+    let validity = if let Some(validity) = values.validity() {
+        let validity = take_bitmap_unchecked(validity, indices.values().as_slice());
+        combine_validities_and(Some(&validity), indices.validity())
+    } else {
+        indices.validity().cloned()
+    };
 
-            {
-                offsets.as_slice().windows(2).enumerate().for_each(
-                    |(i, window): (usize, &[i64])| {
-                        if window[0] == window[1] {
-                            // offsets are equal, slot is null
-                            unset_bit_raw(validity_ptr, i);
-                        }
-                    },
-                );
-            }
-            Some(validity.into())
-        } else {
-            None
-        };
     let dtype = ListArray::<i64>::default_datatype(taken.data_type().clone());
     // Safety:
     // offsets are monotonically increasing
