+++ polars/polars-core/src/frame/hash_join/zip_outer.rs
@@ -1,0 +1,108 @@
+use super::*;
+
+pub trait ZipOuterJoinColumn {
+    fn zip_outer_join_column(
+        &self,
+        _right_column: &Series,
+        _opt_join_tuples: &[(Option<IdxSize>, Option<IdxSize>)],
+    ) -> Series {
+        unimplemented!()
+    }
+}
+
+impl<T> ZipOuterJoinColumn for ChunkedArray<T>
+where
+    T: PolarsIntegerType,
+    ChunkedArray<T>: IntoSeries,
+{
+    fn zip_outer_join_column(
+        &self,
+        right_column: &Series,
+        opt_join_tuples: &[(Option<IdxSize>, Option<IdxSize>)],
+    ) -> Series {
+        let right_ca = self.unpack_series_matching_type(right_column).unwrap();
+
+        let left_rand_access = self.take_rand();
+        let right_rand_access = right_ca.take_rand();
+
+        opt_join_tuples
+            .iter()
+            .map(|(opt_left_idx, opt_right_idx)| {
+                if let Some(left_idx) = opt_left_idx {
+                    unsafe { left_rand_access.get_unchecked(*left_idx as usize) }
+                } else {
+                    unsafe {
+                        let right_idx = opt_right_idx.unwrap_unchecked();
+                        right_rand_access.get_unchecked(right_idx as usize)
+                    }
+                }
+            })
+            .collect_trusted::<ChunkedArray<T>>()
+            .into_series()
+    }
+}
+
+macro_rules! impl_zip_outer_join {
+    ($chunkedtype:ident) => {
+        impl ZipOuterJoinColumn for $chunkedtype {
+            fn zip_outer_join_column(
+                &self,
+                right_column: &Series,
+                opt_join_tuples: &[(Option<IdxSize>, Option<IdxSize>)],
+            ) -> Series {
+                let right_ca = self.unpack_series_matching_type(right_column).unwrap();
+
+                let left_rand_access = self.take_rand();
+                let right_rand_access = right_ca.take_rand();
+
+                opt_join_tuples
+                    .iter()
+                    .map(|(opt_left_idx, opt_right_idx)| {
+                        if let Some(left_idx) = opt_left_idx {
+                            unsafe { left_rand_access.get_unchecked(*left_idx as usize) }
+                        } else {
+                            unsafe {
+                                let right_idx = opt_right_idx.unwrap_unchecked();
+                                right_rand_access.get_unchecked(right_idx as usize)
+                            }
+                        }
+                    })
+                    .collect::<$chunkedtype>()
+                    .into_series()
+            }
+        }
+    };
+}
+impl_zip_outer_join!(BooleanChunked);
+impl_zip_outer_join!(Utf8Chunked);
+impl_zip_outer_join!(BinaryChunked);
+
+impl ZipOuterJoinColumn for Float32Chunked {
+    fn zip_outer_join_column(
+        &self,
+        right_column: &Series,
+        opt_join_tuples: &[(Option<IdxSize>, Option<IdxSize>)],
+    ) -> Series {
+        self.apply_as_ints(|s| {
+            s.zip_outer_join_column(
+                &right_column.bit_repr_small().into_series(),
+                opt_join_tuples,
+            )
+        })
+    }
+}
+
+impl ZipOuterJoinColumn for Float64Chunked {
+    fn zip_outer_join_column(
+        &self,
+        right_column: &Series,
+        opt_join_tuples: &[(Option<IdxSize>, Option<IdxSize>)],
+    ) -> Series {
+        self.apply_as_ints(|s| {
+            s.zip_outer_join_column(
+                &right_column.bit_repr_large().into_series(),
+                opt_join_tuples,
+            )
+        })
+    }
+}
