+++ polars/tests/it/core/mod.rs
@@ -2,6 +2,7 @@ mod date_like;
 mod groupby;
 mod joins;
 mod list;
+mod ops;
 #[cfg(feature = "pivot")]
 mod pivot;
 #[cfg(feature = "rolling_window")]
